moved to new dir and seperated dialog to its own component
