+++ client/src/pages/chat/dashboard/chat-dashboard.jsx
@@ -9,9 +9,7 @@ import { Card, CardContent, CardFooter, CardHeader, CardTitle } from "@/componen
 import { EmptyItems } from "@/components/empty-items";
 import { Button } from "@/components/ui/button";
 import { KebabMenu } from "@/components/ui/kebab-menu";
-import { Input } from "@/components/ui/input";
-import { Dialog, DialogContent, DialogHeader, DialogTitle, DialogFooter } from "@/components/ui/dialog";
-import { AlertDialog, AlertDialogAction, AlertDialogCancel, AlertDialogContent, AlertDialogDescription, AlertDialogFooter, AlertDialogHeader, AlertDialogTitle } from "@/components/ui/alert-dialog";
+import { ChatDialogs } from "./chat-dialogs";
 import { Plus, Clock, Pencil, Trash2, Loader2 } from "lucide-react";
 
 const ChatDashboard = () => {
@@ -20,7 +18,6 @@ const ChatDashboard = () => {
     // State for modals
     const [renameDialog, setRenameDialog] = useState({ open: false, chatId: null, currentTitle: '' });
     const [deleteDialog, setDeleteDialog] = useState({ open: false, chatId: null, chatTitle: '' });
-    const [newTitle, setNewTitle] = useState('');
 
     const {
         chats,
@@ -51,34 +48,15 @@ const ChatDashboard = () => {
         navigate('/chat/new-chat', { state: { chatId } });
     };
 
-    // Rename chat handlers
+    // Dialog handlers
     const handleRenameClick = (chat) => {
         setRenameDialog({
             open: true,
             chatId: chat.id,
             currentTitle: chat.title
         });
-        setNewTitle(chat.title);
     };
 
-    const handleRenameSubmit = async () => {
-        if (!newTitle.trim() || !renameDialog.chatId) return;
-
-        try {
-            await renameChat(renameDialog.chatId, newTitle.trim());
-            setRenameDialog({ open: false, chatId: null, currentTitle: '' });
-            setNewTitle('');
-        } catch (error) {
-            console.error('Error renaming chat:', error);
-        }
-    };
-
-    const handleRenameCancel = () => {
-        setRenameDialog({ open: false, chatId: null, currentTitle: '' });
-        setNewTitle('');
-    };
-
-    // Delete chat handlers
     const handleDeleteClick = (chat) => {
         setDeleteDialog({
             open: true,
@@ -87,20 +65,17 @@ const ChatDashboard = () => {
         });
     };
 
-    const handleDeleteConfirm = async () => {
-        if (!deleteDialog.chatId) return;
-
-        try {
-            await deleteChat(deleteDialog.chatId);
-            setDeleteDialog({ open: false, chatId: null, chatTitle: '' });
-        } catch (error) {
-            console.error('Error deleting chat:', error);
-        }
-    };
-
-    const handleDeleteCancel = () => {
-        setDeleteDialog({ open: false, chatId: null, chatTitle: '' });
-    };
+    // Error display component
+    const ErrorDisplay = ({ error, title, onRetry }) => (
+        <div className="text-red-500 text-sm mt-2">
+            {title}: {error.message}
+            {onRetry && (
+                <Button variant="outline" size="sm" onClick={onRetry} className="ml-2">
+                    Retry
+                </Button>
+            )}
+        </div>
+    );
 
     if (chatsError) {
         return (
@@ -135,19 +110,17 @@ const ChatDashboard = () => {
 
             {/* Error messages */}
             {createChatError && (
-                <div className="text-red-500 text-sm mt-2">
-                    Error creating chat: {createChatError.message}
-                </div>
+                <ErrorDisplay
+                    error={createChatError}
+                    title="Error creating chat"
+                    onRetry={handleNewChat}
+                />
             )}
             {renameChatError && (
-                <div className="text-red-500 text-sm mt-2">
-                    Error renaming chat: {renameChatError.message}
-                </div>
+                <ErrorDisplay error={renameChatError} title="Error renaming chat" />
             )}
             {deleteChatError && (
-                <div className="text-red-500 text-sm mt-2">
-                    Error deleting chat: {deleteChatError.message}
-                </div>
+                <ErrorDisplay error={deleteChatError} title="Error deleting chat" />
             )}
 
             <div className="flex justify-between gap-3">
@@ -185,74 +158,24 @@ const ChatDashboard = () => {
                 </div>
             )}
 
-            {/* Rename Dialog */}
-            <Dialog open={renameDialog.open} onOpenChange={(open) => !open && handleRenameCancel()}>
-                <DialogContent aria-describedby="rename-dialog-description">
-                    <DialogHeader>
-                        <DialogTitle>Rename Chat</DialogTitle>
-                    </DialogHeader>
-                    <div className="py-4">
-                        <p id="rename-dialog-description" className="text-sm text-muted-foreground mb-3">
-                            Enter a new title for your chat conversation.
-                        </p>
-                        <Input
-                            value={newTitle}
-                            onChange={(e) => setNewTitle(e.target.value)}
-                            placeholder="Enter new chat title"
-                            onKeyDown={(e) => {
-                                if (e.key === 'Enter') {
-                                    handleRenameSubmit();
-                                }
-                            }}
-                        />
-                    </div>
-                    <DialogFooter>
-                        <Button variant="outline" onClick={handleRenameCancel}>
-                            Cancel
-                        </Button>
-                        <Button
-                            onClick={handleRenameSubmit}
-                            disabled={isRenamingChat || !newTitle.trim()}
-                        >
-                            {isRenamingChat ? (
-                                <Loader2 className="h-4 w-4 animate-spin mr-2" />
-                            ) : null}
-                            Rename
-                        </Button>
-                    </DialogFooter>
-                </DialogContent>
-            </Dialog>
-
-            {/* Delete Confirmation Dialog */}
-            <AlertDialog open={deleteDialog.open} onOpenChange={(open) => !open && handleDeleteCancel()}>
-                <AlertDialogContent>
-                    <AlertDialogHeader>
-                        <AlertDialogTitle>Delete Chat</AlertDialogTitle>
-                        <AlertDialogDescription>
-                            Are you sure you want to delete "{deleteDialog.chatTitle}"? This action cannot be undone.
-                        </AlertDialogDescription>
-                    </AlertDialogHeader>
-                    <AlertDialogFooter>
-                        <AlertDialogCancel onClick={handleDeleteCancel}>
-                            Cancel
-                        </AlertDialogCancel>
-                        <AlertDialogAction
-                            onClick={handleDeleteConfirm}
-                            disabled={isDeletingChat}
-                            className="bg-red-600 hover:bg-red-700"
-                        >
-                            {isDeletingChat ? (
-                                <Loader2 className="h-4 w-4 animate-spin mr-2" />
-                            ) : null}
-                            Delete
-                        </AlertDialogAction>
-                    </AlertDialogFooter>
-                </AlertDialogContent>
-            </AlertDialog>
+            {/* Dialogs */}
+            <ChatDialogs
+                renameDialog={renameDialog}
+                deleteDialog={deleteDialog}
+                onRenameDialogChange={setRenameDialog}
+                onDeleteDialogChange={setDeleteDialog}
+                onRenameSubmit={renameChat}
+                onDeleteConfirm={deleteChat}
+                isRenamingChat={isRenamingChat}
+                isDeletingChat={isDeletingChat}
+            />
         </MainLayout>
     );
 };
 
+/**
+ * Utility function to format time ago
+ */
 const getTimeAgo = (dateString) => {
     const now = new Date();
     const updatedAt = new Date(dateString);
@@ -269,13 +192,16 @@ const getTimeAgo = (dateString) => {
     return updatedAt.toLocaleDateString();
 };
 
+/**
+ * Individual chat card component
+ */
 const ChatCard = ({ chat, onClick, onRename, onDelete, isDeleting }) => {
     const menuItems = [
         {
             icon: Pencil,
             label: "Rename",
             action: (e) => {
-                e?.stopPropagation?.(); // Stop event propagation
+                e?.stopPropagation?.();
                 onRename();
             }
         },
@@ -283,7 +209,7 @@ const ChatCard = ({ chat, onClick, onRename, onDelete, isDeleting }) => {
             icon: Trash2,
             label: "Delete",
             action: (e) => {
-                e?.stopPropagation?.(); // Stop event propagation
+                e?.stopPropagation?.();
                 onDelete();
             },
             disabled: isDeleting
@@ -291,7 +217,7 @@ const ChatCard = ({ chat, onClick, onRename, onDelete, isDeleting }) => {
     ];
 
     const handleMenuClick = (e) => {
-        e.stopPropagation(); // Prevent card click when menu is opened
+        e.stopPropagation();
     };
 
     return (
